Add files via upload
The data is related with direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be ('yes') or not ('no') subscribed. The main objective of the project is as below
1)The features of a customer that affects his/her response towards the marketing campaign.
2)To Predict whether a customer is going to subscribe to the term deposit.